auto: deploy formularacing-0.2
Changes:
M	src/main/java/dev/EfraGroup/formulaRacing/Listener/RegionListener.java
M	src/main/java/dev/EfraGroup/formulaRacing/Utils/RaceActionBarManager.java
M	src/main/java/dev/EfraGroup/formulaRacing/Utils/ScoreboardTimeTrialUtils.java

Stats:
 3 files changed, 6 insertions(+), 8 deletions(-)

+++ src/main/java/dev/EfraGroup/formulaRacing/Listener/RegionListener.java
@@ -760,7 +760,7 @@ public class RegionListener implements Listener {
     private void handleRaceLapCrossing(Player player, Driver driver, Heats heat, Location from, Location to, DatabaseManager.RegionData regionData) {
         UUID uuid = player.getUniqueId();
         long now = System.currentTimeMillis();
-        Long lastCross = (Long)this.lastStartEndCross.get(uuid);
+        Long lastCross = this.lastStartEndCross.get(uuid);
         if (lastCross == null || now - lastCross >= 2000L) {
             this.lastStartEndCross.put(uuid, now);
             if (!driver.isFinished() && !driver.isDnf()) {
